+++ src/Pages/Contact.js
@@ -1,7 +1,17 @@
-import React from 'react'
+import ContactForm from "../Components/Contact/ContactForm";
+import TextContent from "../Components/Contact/TextContent";
+import '../Components/Contact/ContactStyles.css'
 
-const Contact = () => {
-	return <div>¡Contactate con Nosotros!</div>
+function Contact() {
+    return (
+        <div>
+            <h1>¡Contacte con nosotros!</h1>
+            <div className="contact-container">
+                <TextContent />
+                <ContactForm />
+            </div>
+        </div>
+    )
 }
 
 export default Contact
